requirements.txt
# User accounts preparation
django==5.2.4 
djangorestframework==3.16.0
markdown==3.8.2
django-filter==25.1
djangorestframework-simplejwt==5.5.0 
pyjwt==2.9.0
django-cors-headers==4.7.0
dotenv==0.9.9 
python-dotenv==1.1.1
psycopg==3.2.9
psycopg2==2.9.10
dj-rest-auth[with_social]==7.0.1
django-allauth==65.10.0
requests==2.32.4
cryptography==45.0.5

# Documentation
drf-spectacular==0.28.0

